+++ flask_unchained/forms/fields.py
@@ -1,0 +1,2 @@
+from wtforms.fields import *
+from wtforms.fields.html5 import *